LU-4315 doc: add man pages to checkpatch
Add formatting rules for Lustre Manual pages to checkpatch.
This will encourage not only a uniform formatting for new manual
pages, but it will help when updating older manual pages to ensure
they are following new formatting style.

Note that there is a new AVAILABILITY section that is required for
all pages. This section should include the release and commit
versions that the subject of the page was added.
The checkpatch-man.pl script will generate version based on
what appear in the NAME section of the manual which is also
a required section. This only occurs if the AVAILABILITY
section already exists but is missing either "release X.X.X"
or "commit X.X.X-*"

Using the following pages as guidelines for creating style
rules for man pages:
https://liw.fi/manpages/
https://man7.org/linux/man-pages/man7/man-pages.7.html

Test-Parameters: trivial
Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: I5bd7af0cdbeb49a9edc07d0cc2f8a0dccca39a9e
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55954
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>