LU-17928 build: fix 'or' clause for rpmbuild
This fixes the 'or' clause testing for older distro
versions, which was intended to provide backwards
compatibility for builds on distros that didn't natively
support the rpmbuild 'or' keyword. Instead, it always
evaluated to true, making it so that the DKMS package
was never included as a requirement satisfaction.

This change uses the 'or dkms' Requires, if any of the
distro versions are above a certain level.

Test-Parameters: trivial
HPE-bug-id: LUS-12467
Fixes: ab3b29ec39 ("LU-17928 lnet: add kmod devel package")
Signed-off-by: Caleb Carlson <caleb.carlson@hpe.com>
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I529d30a0dfb0b6e2e2bb55834db4c53506992d37
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55877
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: jsimmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>