LU-18047 lnet: sanity-lnet test_260 wrong value
Increased timeout and retries. The test seem to fail randomly at
different parameters, so it doesn't seem to be a configuration
problem. It looks still more like a timeout problem.
Other possibilties to try would be writing the default values in a
file and execute the sysctl commands in one transaction with sysctl.
An alternative would be trying to flush all buffered IO devices with
sync.

Test-Parameters: trivial testlist=sanity-lnet
Signed-off-by: Frank Sehr <fsehr@whamcloud.com>
Change-Id: I02be748d145a4ec2599c208b31bd0d8242eb6df1
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56025
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Cyril Bordage <cbordage@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>