LU-12515 ofd: allow setting a target to readonly
Add a control to toggle read-only mode for an OFD
to have it reject all mutating commands with -EROFS

This can be used to temporarily set a device to readonly mode
while identifying and correcting a misbehaving client.
As this prevents clients from destaging data it should not
be kept in readonly mode for too long else clients will
eventually run out of kernel memory.

Example:
   lctl set_param obdfilter.lustre-OST0000.readonly=1

Test-Parameters: trivial
Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Change-Id: Ia6658fb58aea17624d5c2ef2528696c4355e7b05
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56304
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>