LU-6142 misc: consolidate license files in LICENSES/
This patch is solely meant to clarify how Lustre is
licensed. It doesn't not alter the licensing itself in
anyway.

Consolidate the scattered license files in the LICENSES/
directory at the root of the Lustre tree. Add missing
license files for BSD-2-Clause and LGPL-2.1. This mimics
the Linux kernel - each file in the tree gets annotated
with SPDX text indicating its license and readers refer
to LICENSES/ and COPYING for details.

The COPYING file now clearly reflects that the Lustre
kernel modules, as a whole, are GPLv2. The remaining
components are governed by their own GPLv2 compatible
licenses annotated in the individual files.

Copying the license-rules.rst from Linux and add a note
that clarifies that these refer explicitly to the kernel
modules.

Test-Parameters: trivial
Signed-off-by: Timothy Day <timday@amazon.com>
Change-Id: Iee64881f6fa5f41e4c426f4ebe941fba7ddd31a9
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56294
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>