LU-10249 utils: Fix lustre_rsync program options
The lustre_rsync program options were incomplete and presented
incorrectly, e.g., the -r option.

This patch improves the program options. It adds a short and long
usage message in accordance to the Lustre documentation replacing the
previous usage message.  Moreover, a check to specify the mdt device
was added as it was missing. Other small changes were added to improve
user interaction.

Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I871a218381b371e61f0b3a258a4773d3e755101e
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56407
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>