LU-18231 obd: change obd_replayable into OBDF_REPLAYABLE
Change the use of the obd_replayable bitfield into a bit in
the obd_flags bitmap and use atomic bit set/clear/test
operations on it.
No logical changes to the code and no changes to locking.

Add obd_replayable and previously-converted bitflags to
spelling.txt to help developers catch changes in old patches.

Test-Parameters: trivial
Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Change-Id: I8efb454b9f00f7fb7c65855a7222d32cee73e031
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56419
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>