LU-18247 nodemap: initialize unused fields on disk
Initialize the unused fields in the on-disk nodemap records to zero.
This will allow them to be used in the future for expanded nodemap
functionality.

It may be that initial usage will also depend on storing a magic
into one of the fields so that the rest of the fields can be trusted.

Test-Parameters: trivial testlist=sanity-sec
Test-Parameters: testlist=sanity-quota
Test-Parameters: testlist=sanity-selinux
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I58f46396911e28f91e03a287c5b308dcaf3ebbe5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56450
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Maximilian Dilger <mdilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>