LU-18386 utils: add lfs quota --delimiter option
Add a new --delimiter option to the lfs quota command, allowing
users to specify a custom delimiter for the output. This enhances
the flexibility of the quota display, making it easier to parse
the output for scripts and other automated tools.
The default delimiter is a space.

$ lfs quota --delimiter="," /mnt/lustre
Disk quotas for usr root (uid 0):
      Filesystem, kbytes, bquota, blimit, bgrace, ..., igrace
     /mnt/lustre,   4360,      0,      0,      -, ...,      -
Disk quotas for grp root (gid 0):
      Filesystem, kbytes, bquota, blimit, bgrace, ..., igrace
     /mnt/lustre,   4360,      0,      0,      -, ...,      -

Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: I58faffeeeecf07ae8630d3f518306c5dcec39867
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56739
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Etienne AUJAMES <eaujames@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>