LU-14520 ldlm: rename LCK_MAXMODE/LDLM_MAX_TYPE
Rename the "MAX" definitions for the ldlm_type and ldlm_mode enums,
since they do not actually represent the "maximum" value of the enum.
They are also better named following the enum, while we are at it.

    LCK_MINMODE -> LCK_MODE_MIN
    LCK_MAXMODE -> LCK_MODE_END
    LDLM_MIN_TYPE -> LDLM_TYPE_MIN
    LDLM_MAX_TYPE -> LDLM_TYPE_END

Add spelling.txt rules to notify in-flight patches of this change, but
include backward compatibility macros since they are in lustre_idl.h.

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ia709775ffc60092883971853470195678b2b9b08
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56760
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>