LU-17712 tests: allow multiop to return 0 in recovery-small/157
This patch allows multiop to return 0 in recovery-small/157,
which is conserdered as success because the mmaped I/O
did not hang after the client was evicted.

The patch also skips recovery-small/157 for older MDS since
the test depends on changes made to the MDS code on the server.

Test-Parameters: trivial testlist=recovery-small serverversion=2.15.5

Fixes: 71f8e5d650 ("LU-14708 ptlrpc: skip unnecessary client eviction")
Change-Id: I19d0362ffbccb44c81ec2e09d9b4eccba40b9dcf
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Signed-off-by: Jian Yu <yujian@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56884
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Emoly Liu <emoly@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>